src/main/java/com/example/course_work2/service/impl/ExaminerServiceImpl.java
package com.example.course_work2.service.impl;

import com.example.course_work2.exception.IncorrectQuestionAmountException;
import com.example.course_work2.model.Question;
import com.example.course_work2.service.ExaminerService;
import com.example.course_work2.service.QuestionService;
import org.springframework.stereotype.Service;

import java.util.Collection;
import java.util.HashSet;
import java.util.Set;

@Service
public class ExaminerServiceImpl implements ExaminerService {
    private QuestionService questionService;

    public ExaminerServiceImpl(QuestionService questionService) {
        this.questionService = questionService;
    }

    @Override
    public Collection<Question> getQuestions(int amount) {
        if (amount <= 0 || amount > questionService.getAll().size()) {
            throw new IncorrectQuestionAmountException();
        }
        Set<Question> questions = new HashSet<>();

        while (questions.size() < amount) {
            questions.add(questionService.getRandomQuestion());
        }
        return questions;
    }

}
